+++ test/functional/kernel/basic-fision-fussion-loop/CMakeLists.txt
@@ -10,7 +10,7 @@
 # 
 # Note: KERNEL_BACKENDS is defined in ../CMakeLists.txt
 #
-foreach( BACKEND ${KERNEL_BACKENDS} )
+foreach( BACKEND ${KERNEL_BACKENDS} )                      
       configure_file( test-kernel-basic-fision-fussion-loop.cpp.in
                       test-kernel-basic-fision-fussion-loop-${BACKEND}.cpp )
       raja_add_test( NAME test-kernel-basic-fision-fussion-loop-${BACKEND}
