--- test/functional/kernel/basic-fision-fussion-loop/test-kernel-basic-fision-fussion-loop-segments.hpp
@@ -1,65 +1,0 @@
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
-// Copyright (c) 2016-21, Lawrence Livermore National Security, LLC
-// and RAJA project contributors. See the RAJA/LICENSE file for details.
-//
-// SPDX-License-Identifier: (BSD-3-Clause)
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
-
-#ifndef __TEST_KERNEL_BASIC_FISION_FUSSION_LOOP_HPP__
-#define __TEST_KERNEL_BASIC_FISION_FUSSION_LOOP_HPP__
-
-#include "basic-single-icount-loop-impl.hpp"
-
-TYPED_TEST_SUITE_P(KernelBasicFisionFussionLoopTest);
-template <typename T>
-class KernelBasicFisionFussionLoopTest : public ::testing::Test
-{
-};
-
-TYPED_TEST_P(KernelBasicFisionFussionLoopTest, BasicFisionFussionLoopSegmentKernel)
-{
-  using IDX_TYPE    = typename camp::at<TypeParam, camp::num<0>>::type;
-  using WORKING_RES = typename camp::at<TypeParam, camp::num<1>>::type;
-  using EXEC_POLICY = typename camp::at<TypeParam, camp::num<2>>::type;
-
-  WORKING_RES working_res{WORKING_RES::get_default()};
-  camp::resources::Resource erased_working_res{working_res};
-
-  std::vector<IDX_TYPE> seg_idx;
-
-  // Range segment tests
-  RAJA::TypedRangeSegment<IDX_TYPE> r1( 0, 37 );
-  RAJA::getIndices(seg_idx, r1);
-
-  KernelBasicFisionFussionLoopTestImpl<IDX_TYPE, EXEC_POLICY, WORKING_RES,
-                                      RAJA::TypedRangeSegment<pTestImpl<IDX_TYPE, EXEC_POLICY, WORKING_RES,
-                                                                        RAJA::TypedRangeStrideSegment<IDX_TYPE>>(
-                                                                                                                 rs3, seg_idx, working_res, erased_working_res);
-
-  // List segment tests
-  seg_idx.clear();
-  IDX_TYPE last = IDX_TYPE(10567);
-  srand( time(NULL) );
-  for (IDX_TYPE i = IDX_TYPE(0); i < last; ++i) {
-    IDX_TYPE randval = IDX_TYPE( rand() % RAJA::stripIndexType(last) );
-    if ( i < randval ) {
-      seg_idx.push_back(i);
-    }
-  }
-  RAJA::TypedListSegment<IDX_TYPE> l1( &seg_idx[0], seg_idx.size(), erased_working_res);
-  KernelBasicFisionFussionLoopTestImpl<IDX_TYPE, EXEC_POLICY, WORKING_RES,
-                                      RAJA::TypedListSegment<IDX_TYPE>>(
-                                                                        l1, seg_idx, working_res, erased_working_res);
-
-  // test zero-length list segment
-  seg_idx.clear();
-  RAJA::TypedListSegment<IDX_TYPE> l2( nullptr, seg_idx.size(), erased_working_res);
-  KernelBasicFisionFussionLoopTestImpl<IDX_TYPE, EXEC_POLICY, WORKING_RES,
-                                      RAJA::TypedListSegment<IDX_TYPE>>(
-                                                                        l2, seg_idx, working_res, erased_working_res);
-}
-
-REGISTER_TYPED_TEST_SUITE_P(KernelBasicFisionFussionLoopTest,
-                            BasicFisionFussionLoopSegmentKernel);
-
-#endif  // __TEST_KERNEL_BASIC_FISION_FUSSION_LOOP_HPP__
