+++ test/functional/kernel/basic-fision-fussion-loop/test-kernel-basic-fision-fussion-loop.cpp.in
@@ -201,14 +201,14 @@ camp::list<
 //
 // Cartesian product of types used in parameterized tests
 //
-using @BACKEND@KernelBasicSingleFisionFussionLoopTypes =
+using @BACKEND@KernelBasicFisionFussionLoopTypes =
   Test< camp::cartesian_product<StrongIdxTypeList,
                                 @BACKEND@ResourceList,
                                 @BACKEND@KernelExecPols>>::Types;
 
-1//
+//
 // Instantiate parameterized tests
 //
 INSTANTIATE_TYPED_TEST_SUITE_P(@BACKEND@,
-                               KernelBasicFissionFussionLoopTest,
+                               KernelBasicFisionFussionLoopTest,
                                @BACKEND@KernelBasicFisionFussionLoopTypes);
